==> src/4_statistical_Tests.ipynb <==
# Statistical Tests

import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
from scipy.stats import pearsonr
from scipy import stats

## Histogram functions

def rgb_hist(net, limit): 
    plt.rcParams.update({'font.size': 16})
    n, bins, patches = plt.hist(x=df_e, bins='auto', color='#607c8e',
                                alpha=0.7, rwidth=0.85)
    plt.grid(axis='y', alpha=0.75)
    plt.xlabel('absolut angular estimation error')
    plt.ylabel('frequency')
    plt.title('Histogram of the absolut angular estimation error of the \n ' + net + ' trained on RGB images')
    m = round(np.mean(df_e, axis = 0), 2)
    s = round(np.std(df_e), 2)
    pos = 0.8 * limit
    plt.text(pos, 470, r'$\mu={}$'.format(m))
    plt.text(pos, 430, r'$\sigma={}$'.format(s))
    plt.axvline(x=m)
    maxfreq = n.max()
    minfreq = n.min()
    # Set a clean upper y-axis limit.
    #plt.ylim(ymax=np.ceil(maxfreq / 10) * 10 if maxfreq % 10 else maxfreq + 10, ymin=np.ceil(minfreq / 10))
    plt.xlim(xmin=0, xmax = limit)
    plt.savefig('..\\output\\Diagrams\\{}_RGB_hist.png'.format(net), dpi=300, bbox_inches='tight')

def rgbd_hist(net, limit):
    plt.rcParams.update({'font.size': 16})
    n, bins, patches = plt.hist(x=df_e_d, bins='auto', color='#607c8e',
                                alpha=0.7, rwidth=0.85)
    plt.grid(axis='y', alpha=0.75)
    plt.xlabel('absolut angular estimation error')
    plt.ylabel('frequency')
    plt.title('Histogram of the absolut angular estimation error of the \n ' + net + ' trained on RGBD images')
    m = round(np.mean(df_e_d, axis = 0), 2)
    s = round(np.std(df_e_d), 2)
    pos = 0.8 * limit
    plt.text(pos, 470, r'$\mu={}$'.format(m))
    plt.text(pos, 430, r'$\sigma={}$'.format(s))
    plt.axvline(x=m)
    maxfreq = n.max()
    minfreq = n.min()
    # Set a clean upper y-axis limit.
    #plt.ylim(ymax=np.ceil(maxfreq / 10) * 10 if maxfreq % 10 else maxfreq + 10, ymin=np.ceil(minfreq / 10))
    plt.xlim(xmin=0, xmax =limit)
    plt.savefig('..\\output\\Diagrams\\{}_RGBD_hist.png'.format(net), dpi=300, bbox_inches='tight')

### Run RGB

net = 'RESNET' # ALEX vs VGG16 vs RESNET
dataset = '201129_2031'
alpha = 0.01

add =''

if (net == 'RESNET'):
    limit = 10
elif (net == 'VGG16'):
    limit = 10
elif (net == 'ALEX'):
    limit = 30  

# RGB
csv_rgb = '..\\output\\{}_{}_rgb{}\\Evaluation_SYNTH\\Prognosen.csv'.format(net, dataset, add)
df_rgb = pd.read_csv(csv_rgb)

# RGBD
csv_rgbd = '..\\output\\{}_{}_rgba{}\\Evaluation_SYNTH\\Prognosen.csv'.format(net, dataset, add)
df_rgbd = pd.read_csv(csv_rgbd)

df_rgb.head()


df_e = df_rgb['dot_angular_err_elevation']
rgb_hist(net, limit)

df_e_d = df_rgbd['dot_angular_err_elevation']
rgbd_hist(net, limit)

## T-Statistic

#https://www.hackdeploy.com/python-t-test-a-friendly-guide/
tStat, pValue =  stats.ttest_rel(df_e, df_e_d)
print("P-Value:{0} T-Statistic:{1}".format(pValue,tStat)) 

# interpret via p-value
if pValue > alpha:
    print('Accept null hypothesis that the means are equal.')
else:
    print('Reject the null hypothesis that the means are equal.')
    
print('r = {}'.format(f_r(tStat)))

## Wilcoxon signed-rank test

# https://www.methodenberatung.uzh.ch/de/datenanalyse_spss/unterschiede/zentral/wilkoxon.html
W, p_v = stats.wilcoxon(df_e, df_e_d)
print(W)
print('p = {}'.format(p_v))

def f_z(W, u, o):
    return (W - u) / o

def f_u(n):
    return (n*(n+1))/4
def f_o(n):
    return np.sqrt((n*(n+1)*(2*n+1))/24)

def f_r_w(W, n):
    return abs(f_z(W, f_u(n), f_o(n)) / np.sqrt(n))
          
n = df_e.size
r_w = f_r_w(W, n)
print('r = {}'.format(r_w))

r_r, p_r = stats.pearsonr(df_e, df_e_d)
print(r_r)
print(p_r)
